refactor: align file names and tests with coding conventions
The template documents kebab-case file names and test() over describe/it,
but a few existing files predated that rule. Fix them so the template
practices its own conventions (important now that it doubles as a
learning resource).

- rename Message.ts, GetHelloMessage.ts, HelloResponse.ts and their test
  to kebab-case (class/type names unchanged); update imports
- hello.test.ts: drop describe() wrapper, use test() directly
