+++ __tests__/hello.test.ts
@@ -1,38 +1,33 @@
-import { describe, test, expect } from 'vitest'
+import { test, expect } from 'vitest'
 import { APIGatewayProxyEvent, Context } from 'aws-lambda'
 import { handler } from '../src/handlers/hello'
 
-describe('hello handler', () => {
-  const mockContext: Context = {
-    awsRequestId: 'test-request-id',
-  } as Context
+const mockContext: Context = {
+  awsRequestId: 'test-request-id',
+} as Context
 
-  const mockEvent = {} as APIGatewayProxyEvent
+const mockEvent = {} as APIGatewayProxyEvent
 
-  test('should return 200 status code and correct body', async () => {
-    const result = await handler(mockEvent, mockContext)
+test('hello handler は 200 と正しいボディを返す', async () => {
+  const result = await handler(mockEvent, mockContext)
 
-    expect(result.statusCode).toBe(200)
-    
-    const body = JSON.parse(result.body)
-    expect(body).toHaveProperty(
-      'message',
-      'Hello World from Serverless REST API!'
-    )
-  })
+  expect(result.statusCode).toBe(200)
 
-  test('should handle errors gracefully', async () => {
-    // Mock an error by passing invalid context
-    const invalidContext = null as unknown as Context
+  const body = JSON.parse(result.body)
+  expect(body).toHaveProperty(
+    'message',
+    'Hello World from Serverless REST API!'
+  )
+})
 
-    const result = await handler(mockEvent, invalidContext)
+test('hello handler はエラー時に 500 を返す', async () => {
+  // 不正な context を渡してエラーを発生させる
+  const invalidContext = null as unknown as Context
 
-    expect(result.statusCode).toBe(500)
-    
-    const body = JSON.parse(result.body)
-    expect(body).toHaveProperty(
-      'error',
-      'Internal Server Error'
-    )
-  })
+  const result = await handler(mockEvent, invalidContext)
+
+  expect(result.statusCode).toBe(500)
+
+  const body = JSON.parse(result.body)
+  expect(body).toHaveProperty('error', 'Internal Server Error')
 })
