+++ src/domain/entities/message.ts
